audio-service/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyYAML==6.0.1
psutil==5.9.5
